requirements.txt
# ─── Core ─────────────────────────────────────────────────────
python-dotenv==1.0.1

# ─── Database ─────────────────────────────────────────────────
mysql-connector-python==8.3.0

# ─── LLM - Ollama client ──────────────────────────────────────
ollama==0.2.1
requests>=2.28

# ─── AirLLM - Layer-wise inference ────────────────────────────
airllm==2.11.0
transformers==4.41.2
accelerate==0.30.1
bitsandbytes==0.43.1

# ─── PyTorch (CUDA 12.1 - RTX 2050) ──────────────────────────
--extra-index-url https://download.pytorch.org/whl/cu121
torch==2.3.0+cu121
torchvision==0.18.0+cu121
torchaudio==2.3.0+cu121

# ─── Embeddings ───────────────────────────────────────────────
sentence-transformers==3.0.1

# ─── Vector Store ─────────────────────────────────────────────
chromadb==0.5.0

# ─── Document Loaders ─────────────────────────────────────────
PyMuPDF==1.24.5          # PDF → fitz
python-docx==1.1.2       # DOCX
pandas==2.2.2            # CSV

# ─── LangChain (utility) ──────────────────────────────────────
langchain==0.2.5
langchain-community==0.2.5

# ─── UI ───────────────────────────────────────────────────────
streamlit==1.35.0

# ─── Utilities ────────────────────────────────────────────────
numpy==1.26.4
tqdm==4.66.4
loguru==0.7.2

# ─── Knowledge Base / Dataset Collection ──────────────────────
datasets==2.20.0          # HuggingFace datasets
huggingface-hub==0.23.4   # HF Hub access